requirements.txt
asgiref==3.8.0
Django==3.2.20
django-bootstrap5==22.2
django-debug-toolbar==4.2.0
django-taggit==4.0.0
djangorestframework==3.15.0
flake8==5.0.4
gunicorn==21.2.0
mccabe==0.7.0
packaging==24.0
Pillow==10.1.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
